src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "boto3>=1.40.64",
    "python-multipart>=0.0.9",
    "fastapi>=0.120.4",
    "httpx>=0.28.1",
    "openai>=1.40.0",
    "pypdf>=4.3.1",
    "langchain>=1.0.3",
    "langgraph>=1.0.2",
    "langsmith>=0.4.39",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn[standard]>=0.38.0",
    "kernel>=0.0.1",
]
